requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
astroid==2.5.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.7
django-crum==0.7.9
django-rest-framework==0.1.0
django-userforeignkey==0.3.0
djangorestframework==3.12.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.1.2
psycopg2==2.8.6
pylint==2.7.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
pytz==2021.1
reportlab==3.5.65
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
